requirements.txt
fastapi
sqlalchemy
strawberry-graphql[debug-server]
